stableswap: address review comments on PR #2839
Hardfork gating (comment on liquidity_pool_evaluator.cpp:51)
  The gate tested the VALUE of the extension, so pool_type = constant_product, the
  default, passed before the hardfork -- and an operation carrying a set-but-default
  extension serialises differently from one carrying none, whatever the value means.
  All three evaluators now test presence, and do it first, before anything is read
  or computed. proposal_operation_hardfork_visitor gets the same checks, since a
  proposal is validated when it is created rather than when it executes.

Withdrawal fee (comment on liquidity_pool_evaluator.cpp:449)
  It was bypassed, as suspected. The single-asset branch set the reported fees to
  zero and returned before withdrawal_fee_percent was ever applied -- that happens
  further down, in the proportional path. Two single-asset withdrawals were a way
  out of the pool that cost nothing. The fee is charged and reported there now. The
  imbalance fee is deliberately not reported alongside it: it is already priced into
  the payout by being left in the pool before solving for what the shares are worth,
  and reporting it twice would overstate what the withdrawal cost.

Smaller points
  - Comments in English (liquidity_pool_evaluator.cpp:311, :316, :511).
  - SS_MAX_ITER and the two loop counters are int16_t rather than int (:73).
  - The deposit and withdraw extensions no longer say "Unused. Reserved for future
    use." -- they are used, and now say what they hold (:146, :206).
  - Dropped the two FC_REFLECT_TYPENAME declarations for the extension types (:263).
  - build-docker.yml reverted to what the branch found it as (:5).

Still open from the review: market trading fees on deposit and withdraw (:283,
:449), moving the validate()-only assertions out of the evaluator (:64), the
avoidable recomputations (:244, :284, :402), explicit time-based gating (:389), and
the rounding question on Newton's method (:143).

+++ libraries/chain/include/graphene/chain/stableswap.hpp
@@ -70,7 +70,7 @@ namespace stableswap {
 /// Number of coins in the pool. Fixed at 2 for BitShares liquidity pools.
 constexpr uint32_t SS_N_COINS = 2;
 /// Maximum number of Newton iterations before we give up converging.
-constexpr int SS_MAX_ITER = 255;
+constexpr int16_t SS_MAX_ITER = 255;
 
 namespace detail {
 
@@ -134,7 +134,7 @@ inline fc::uint128_t compute_d( const fc::uint128_t& x, const fc::uint128_t& y, 
    wide_uint recent[SS_CYCLE_WINDOW] = {};
    size_t filled = 0;
 
-   for( int i = 0; i < SS_MAX_ITER; ++i )
+   for( int16_t i = 0; i < SS_MAX_ITER; ++i )
    {
       // D_P = D^(n+1) / (n^n * prod(x_i)) ; for n=2: D_P = D^3 / (4 * x * y)
       // Computed in a 256-bit accumulator: the D_P*D intermediate below can transiently
@@ -213,7 +213,7 @@ inline fc::uint128_t compute_new_y( const fc::uint128_t& new_x, const fc::uint12
    wide_uint y = d256;       // initial guess
    wide_uint y_prev;
 
-   for( int i = 0; i < SS_MAX_ITER; ++i )
+   for( int16_t i = 0; i < SS_MAX_ITER; ++i )
    {
       y_prev = y;
       // y = (y^2 + c) / (2y + b - D)
